Constrain size_t to int to prevent conversion error

+++ wsutil/base32.c
@@ -170,18 +170,21 @@ encode_sequence(const unsigned char *plain, int len, unsigned char *coded)
 	}
 }
 
-size_t
-ws_base32_encode_length(size_t plain_length)
+int
+ws_base32_encode_length(int plain_length)
 {
 	return (plain_length / 5) * 8 + ((plain_length) % 5 ? 8 : 0);
 }
 
 void
-ws_base32_encode(const unsigned char *plain, size_t len, unsigned char *coded)
+ws_base32_encode(const unsigned char *plain, int len, unsigned char *coded)
 {
 	// All the hard work is done in encode_sequence(),
 	// here we just need to feed it the data sequence by sequence.
-	for (size_t i = 0, j = 0; i < len; i += 5, j += 8) {
+	for (int i = 0, j = 0; i < len; i += 5, j += 8) {
+		// The `len` parameter used to be a size_t, but since encode_sequence
+		// can only take an int, I chose to constrain to the smaller size,
+		// just to be on the safe side.
 		encode_sequence(&plain[i], min(len - i, 5), &coded[j]);
 	}
 }
